+++ core/src/renderer/layout/layout_point.rs
@@ -5,28 +5,28 @@
 /// https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/platform/geometry/layout_point.h
 #[derive(Debug, Clone, PartialEq, Copy)]
 pub struct LayoutPoint {
-    x: f64,
-    y: f64,
+    x: i64,
+    y: i64,
 }
 
 impl LayoutPoint {
-    pub fn new(x: f64, y: f64) -> Self {
+    pub fn new(x: i64, y: i64) -> Self {
         Self { x, y }
     }
 
-    pub fn x(&self) -> f64 {
+    pub fn x(&self) -> i64 {
         self.x
     }
 
-    pub fn y(&self) -> f64 {
+    pub fn y(&self) -> i64 {
         self.y
     }
 
-    pub fn set_x(&mut self, x: f64) {
+    pub fn set_x(&mut self, x: i64) {
         self.x = x;
     }
 
-    pub fn set_y(&mut self, y: f64) {
+    pub fn set_y(&mut self, y: i64) {
         self.y = y;
     }
 }
